requirements.txt
fastapi
python-dotenv
requests
pillow
uvicorn